refacto(folios) refactor create main folio and update child
