+++ start/routes.ts
@@ -5,7 +5,6 @@ const RootController = () => import('#controllers/system/root_controller')
 const ApiInfoController = () => import('#controllers/system/api_info_controller')
 const CardsController = () => import('#controllers/product/cards_controller')
 const ScanController = () => import('#controllers/product/scan_controller')
-const FoliosController = () => import('#controllers/product/folios_controller')
 const ChildFoliosController = () => import('#controllers/product/child_folios_controller')
 const MainFoliosController = () => import('#controllers/product/main_folios_controller')
 const CardFoliosController = () => import('#controllers/product/card_folios_controller')
@@ -36,7 +35,7 @@ router
             router.get('/:id', [ChildFoliosController, 'show']).use(middleware.auth())
             router.get('/:id/cards', [ChildFoliosController, 'cards']).use(middleware.auth())
             router.post('/', [ChildFoliosController, 'store']).use(middleware.auth())
-            router.post('/init', [FoliosController, 'init']).use(middleware.auth())
+            router.post('/init', [MainFoliosController, 'store']).use(middleware.auth())
             router.post('/cards', [CardFoliosController, 'collect']).use(middleware.auth())
             router.patch('/cards/:id', [CardFoliosController, 'occurrence']).use(middleware.auth())
             router.delete('/cards/:id', [CardFoliosController, 'delete']).use(middleware.auth())
