+++ tests/functional/controllers/main_folios_controller.spec.ts
@@ -825,4 +825,32 @@ test.group('Main folio controller', (group) => {
     assert.equal(statistics.totalCardsCount, 5) // Should count all occurrences
     assert.equal(statistics.cardMarketPrice, '50.00') // 5 cards × 10.0 each
   })
+
+  test('init - it should create a main folio for a user', async ({ client, assert }) => {
+    const response = await client
+      .post('/api/v1/folios/init')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(200)
+    assert.properties(response.body(), ['message'])
+  })
+
+  test('init - it should not create duplicate main folio', async ({ client, assert }) => {
+    await FolioFactory.merge({
+      userId: '123',
+      isRoot: true,
+    }).create()
+
+    const response = await client
+      .post('/api/v1/folios/init')
+      .header('Authorization', 'Bearer fake-token-for-testing')
+
+    response.assertStatus(409)
+    assert.properties(response.body(), ['code', 'message'])
+  })
+
+  test('init - it should require authentication', async ({ client }) => {
+    const response = await client.post('/api/v1/folios/init')
+    response.assertStatus(401)
+  })
 })
